# doc/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TRO & LAM API
  version: 1.5.0
  description: |
    Contract frozen from doc/planning/04-data-model.md and 05-api-contracts.md.
    All application routes use /api/v1. Health probes keep the scaffold response
    shape. Provider webhook wire payload and ACK are delegated to the current
    provider documentation; browser redirects never mutate payment state.
servers:
  - url: /api/v1
...
paths:
...
  /account/vouchers:
    get:
      operationId: listOwnVouchers
      tags: [account]
      x-actors: [customer, staff, admin]
      responses: { '200': { $ref: '#/components/responses/Success' }, '401': { $ref: '#/components/responses/Error' } }
...
  /admin/vouchers:
    get:
      operationId: listAdminVouchers
      tags: [admin]
      x-actors: [admin]
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' } }
    post:
      operationId: issueCustomerVoucher
      tags: [admin]
      x-actors: [admin]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminVoucherWrite' } } } }
      responses: { '201': { $ref: '#/components/responses/Created' }, '400': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /admin/vouchers/{id}/revoke:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      operationId: revokeCustomerVoucher
      tags: [admin]
      x-actors: [admin]
      responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' } }
...
components:
  parameters:
    Id: { in: path, name: id, required: true, schema: { type: string, minLength: 1 } }
...
  responses:
    Success:
      description: Successful response envelope.
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
    Created:
      description: Resource created.
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
...
    Error:
      description: Standardized application error; health and provider webhooks are exceptions.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  schemas:
    SuccessEnvelope: { $ref: './schemas/dtos.schema.json#/$defs/successEnvelope' }
...
    ErrorEnvelope: { $ref: './schemas/dtos.schema.json#/$defs/errorEnvelope' }
...
    AdminVoucherWrite: { $ref: './schemas/dtos.schema.json#/$defs/adminVoucherWrite' }
...